# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  name: string;
  email: string;
  phone: string;
  service: string;
  /** @nullable */
  doctor?: string | null;
  /** ISO date string */
  date: string;
  time: string;
  /** @nullable */
  message?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}

export interface AppointmentInput {
  name: string;
  email: string;
  phone: string;
  service: string;
  doctor?: string;
  date: string;
  time: string;
  message?: string;
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface Testimonial {
  id: number;
  name: string;
  rating: number;
  review: string;
  treatment: string;
  avatarUrl: string;
  date?: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  author: string;
  category: string;
  imageUrl: string;
  publishedAt: string;
  /** @nullable */
  readTime?: number | null;
}

export interface Faq {
  id: number;
  question: string;
  answer: string;
  category: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  icon: string;
  imageUrl: string;
  price: string;
  /** @nullable */
  duration?: string | null;
  highlights?: string[];
}

export type DoctorSocial = {
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  twitter?: string | null;
};

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  bio: string;
  imageUrl: string;
  education: string;
  experience: number;
  languages?: string[];
  social?: DoctorSocial;
}

export interface NewsletterInput {
  email: string;
  /** @nullable */
  name?: string | null;
}

export interface NewsletterSubscription {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  createdAt: string;
}

export interface ClinicStats {
  patientsServed: number;
  yearsExperience: number;
  doctorsCount: number;
  satisfactionRate: number;
  awardsWon?: number;
}

export type ListBlogPostsParams = {
/**
 * Max number to return
 */
limit?: number;
};