# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dental Clinic API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
          description: Max number to return
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
...
components:
  schemas:
...
    BlogPost:
      type: object
      required: [id, title, slug, excerpt, content, author, category, imageUrl, publishedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        author:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        publishedAt:
          type: string
        readTime:
          type: ["integer", "null"]
...