# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  author: string;
  category: string;
  imageUrl: string;
  publishedAt: string;
  /** @nullable */
  readTime?: number | null;
}